requirements.txt
pyramid-storage==0.0.8
kinto==1.8.0
